:sparkles: feat: 고용주 인재 검색 api 변경사항 반영 #364
